Python-doc/requirements.txt
psycopg2-binary
python-dotenv
bcrypt
PyJWT
fastapi
uvicorn
pydantic[email]
python-multipart